@@ -85,57 +85,54 @@ namespace cpp
           {
             std::string kernel_file = args[0];
 
-            std::shared_ptr<output_stack> os  = this->unit->get_stack();
-            std::shared_ptr<translator>   t   = this->unit->get_translator();
+            output_stack& os = this->unit->get_stack();
+            translator&   t  = this->unit->get_translator();
 
-            if(os != nullptr && t != nullptr)
+            buffer& buf = os.top_buffer();
+            enum process_type type = os.top_process_type();
+
+            // set up a new in-memory buffer to hold the translated kernel
+            buffer kernel_buffer;
+
+            // inherit current decoration (line prefixes, skipping blank lines) from parent buffer
+            kernel_buffer.inherit_decoration(buf);
+
+            // the lines of the kernel have to be quoted with ", so ask buffer to set up new delimiters
+            // also skip blank lines, which [usually] don't help in stringized kernels
+            kernel_buffer.push_delimiter(OUTPUT_VEXCL_KERNEL_LINE_PRE, OUTPUT_VEXCL_KERNEL_LINE_POST);
+            kernel_buffer.push_skip_blank(true);
+
+            // set up a filter to ensure we emit only printable strings.
+            // The bound arguments here are:
+            // * 'false' disables quoting by to_printable())
+            // * 'true' allows newlines to pass through
+            std::function<std::string(const std::string&)> filter = std::bind(to_printable, std::placeholders::_1, false, true);
+
+            // translate into our new in-memory buffer, preserving the type of translation (ie. core or implementation)
+            unsigned int replacements = t.translate(kernel_file, kernel_buffer, type, &filter);
+
+            // merge new buffer into old one
+            buf.merge(kernel_buffer);
+
+            // output kernel name string
+            if(args.size() >= 2)
               {
-                buffer& buf = os->top_buffer();
-                enum process_type type = os->top_process_type();
-
-                // set up a new in-memory buffer to hold the translated kernel
-                buffer kernel_buffer;
-
-		            // inherit current decoration (line prefixes, skipping blank lines) from parent buffer
-		            kernel_buffer.inherit_decoration(buf);
-
-                // the lines of the kernel have to be quoted with ", so ask buffer to set up new delimiters
-                // also skip blank lines, which [usually] don't help in stringized kernels
-                kernel_buffer.push_delimiter(OUTPUT_VEXCL_KERNEL_LINE_PRE, OUTPUT_VEXCL_KERNEL_LINE_POST);
-                kernel_buffer.push_skip_blank(true);
-
-		            // set up a filter to ensure we emit only printable strings.
-		            // The bound arguments here are:
-                // * 'false' disables quoting by to_printable())
-                // * 'true' allows newlines to pass through
-                std::function<std::string(const std::string&)> filter = std::bind(to_printable, std::placeholders::_1, false, true);
-
-		            // translate into our new in-memory buffer, preserving the type of translation (ie. core or implementation)
-                unsigned int replacements = t->translate(kernel_file, kernel_buffer, type, &filter);
-
-		            // merge new buffer into old one
-		            buf.merge(kernel_buffer);
-
-                // output kernel name string
-                if(args.size() >= 2)
-                  {
-                    buf.write_to_end(OUTPUT_VEXCL_KERNEL_PRE + args[1] + OUTPUT_VEXCL_KERNEL_POST);
-                  }
-                else
-                  {
-                    buf.write_to_end(OUTPUT_VEXCL_UNKNOWN_KERNEL);
-                  }
-
-                // output trailing content, which must appear *before* the location of the new pool, if required
-                if(args.size() >= 3)
-                  {
-                    buf.write_to_end(args[2]);
-                    buf.write_to_end("");  // skip a line to avoid collisions with the new pool (set below)
-                  }
-
-                // mark this point as the beginning of a new pool
-                rval = this->printer.comment(OUTPUT_KERNEL_LOCATION);
+                buf.write_to_end(OUTPUT_VEXCL_KERNEL_PRE + args[1] + OUTPUT_VEXCL_KERNEL_POST);
               }
+            else
+              {
+                buf.write_to_end(OUTPUT_VEXCL_UNKNOWN_KERNEL);
+              }
+
+            // output trailing content, which must appear *before* the location of the new pool, if required
+            if(args.size() >= 3)
+              {
+                buf.write_to_end(args[2]);
+                buf.write_to_end("");  // skip a line to avoid collisions with the new pool (set below)
+              }
+
+            // mark this point as the beginning of a new pool
+            rval = this->printer.comment(OUTPUT_KERNEL_LOCATION);
           }
 
         return(rval);
